Mask neutron services
Disable external ways to manage services which are already managed by pacemaker.

Related-Bug: #1652748

+++ deployment/puppet/tweaks/spec/defines/tweaks_ubuntu_service_override_spec.rb
@@ -22,16 +22,23 @@ describe 'tweaks::ubuntu_service_override' do
       it 'configures with the default params' do
         should contain_tweaks__ubuntu_service_override(title)
         if facts[:operatingsystem] == 'Ubuntu'
-          should contain_file('create-policy-rc.d').with(
-            :ensure  => 'present',
-            :path    => '/usr/sbin/policy-rc.d',
-            :content => "#!/bin/bash\nexit 101",
-            :mode    => '0755',
-            :owner   => 'root',
-            :group   => 'root')
-          should contain_exec('remove-policy-rc.d').with(
-            :command => 'rm -f /usr/sbin/policy-rc.d',
-            :onlyif  => 'test -f /usr/sbin/policy-rc.d')
+          if facts[:operatingsystemmajrelease] >= '15.04'
+            file { "/etc/systemd/system/${service_name}.service":
+              ensure => 'link',
+              target => '/dev/null',
+            }
+          else
+            should contain_file('create-policy-rc.d').with(
+              :ensure  => 'present',
+              :path    => '/usr/sbin/policy-rc.d',
+              :content => "#!/bin/bash\nexit 101",
+              :mode    => '0755',
+              :owner   => 'root',
+              :group   => 'root')
+            should contain_exec('remove-policy-rc.d').with(
+              :command => 'rm -f /usr/sbin/policy-rc.d',
+              :onlyif  => 'test -f /usr/sbin/policy-rc.d')
+          end
         else
           should_not contain_file('create-policy-rc.d')
           should_not contain_exec('remove-policy-rc.d')
